refactor: 코드리뷰반영
