+++ src/entities/schedule/models/schedule.query.ts
@@ -3,5 +3,5 @@ export const scheduleQueryKeys = {
   userSchedules: (userId: number | string, start: string, end: string) =>
     [...scheduleQueryKeys.all, `user-${userId}`, `${start}~${end}`] as const,
   shareSchedules: (userId: number | string) => [...scheduleQueryKeys.all, 'share', userId] as const,
-  detailedSchedule: (scheduleId: string) => [...scheduleQueryKeys.all, 'schdule', scheduleId] as const,
+  detailedSchedule: (scheduleId: string) => [...scheduleQueryKeys.all, 'schedule', scheduleId] as const,
 }
