+++ src/entities/schedule/models/use-month-store.tsx
@@ -24,13 +24,14 @@ interface FriendMonthsState {
   setMonths: (friendId: string, updater: SetStateAction<Month[]>) => void
   clearFriend: (friendId: string) => void
 }
+const INITIAL_MONTH: Month[] = getInitialMonth(false)
 
 export const useFriendMonthsStore = create<FriendMonthsState>()((set, get) => ({
   byFriend: {},
-  getMonths: (friendId) => get().byFriend[friendId] ?? getInitialMonth(false),
+  getMonths: (friendId) => get().byFriend[friendId] ?? INITIAL_MONTH,
   setMonths: (friendId, updater) =>
     set((s) => {
-      const prev = s.byFriend[friendId] ?? getInitialMonth(false)
+      const prev = s.byFriend[friendId] ?? INITIAL_MONTH
       const next = typeof updater === 'function' ? (updater as (p: Month[]) => Month[])(prev) : updater
       return { byFriend: { ...s.byFriend, [friendId]: next } }
     }),
@@ -43,10 +44,8 @@ export const useFriendMonthsStore = create<FriendMonthsState>()((set, get) => ({
 }))
 
 export function useFriendMonths(friendId: string) {
-  const getMonths = useFriendMonthsStore((s) => s.getMonths)
+  const months = useFriendMonthsStore((s) => s.getMonths(friendId))
   const setMonthsById = useFriendMonthsStore((s) => s.setMonths)
-  const months = getMonths(friendId)
   const setMonths = (updater: SetStateAction<Month[]>) => setMonthsById(friendId, updater)
-
   return { months, setMonths }
 }
