+++ src/features/calendar/hooks/index.ts
@@ -1,5 +1,5 @@
 export * from './use-calendar-ref'
 export * from './use-infinite-calendar-scroll'
-export * from './use-scroll-to-current-month.ts'
-export * from './use-auto-scroll-to-current-month.ts'
+export * from './use-scroll-to-current-month'
+export * from './use-auto-scroll-to-current-month'
 export * from './use-visible-month-observer'
