+++ src/features/calendar/ui/__stories__/monthly-calendar.stories.tsx
@@ -5,7 +5,7 @@ import { CalendarProvider, MonthlyCalendar } from '..'
 import type { Meta, StoryObj } from '@storybook/react-vite'
 
 const meta: Meta<typeof MonthlyCalendar> = {
-  title: 'shared/calendar/monthly-calendar',
+  title: 'features/calendar/monthly-calendar',
   component: MonthlyCalendar,
   tags: ['autodocs'],
   decorators: [
